feat: display quick links in cmd palette
