+++ apps/desktop/src/lib/stores/index.ts
@@ -3,3 +3,4 @@ export * from "./appState"
 export * from "./winExtMap"
 export * from "./extensions"
 export * from "./auth"
+export * from "./quick-links"
