+++ packages/ui/src/components/main/index.ts
@@ -3,4 +3,5 @@ export { default as CustomCommandInput } from "./CustomCommandInput.svelte"
 export { default as GlobalCommandPaletteFooter } from "./GlobalCommandPaletteFooter.svelte"
 export { default as ExtCmdsGroup } from "./ExtCmdsGroup.svelte"
 export { default as SystemCmds } from "./SystemCmds.svelte"
+export { default as QuickLinks } from "./QuickLinks.svelte"
 export * from "./types"
